corrige logica de recorrecia no calculo de base fee
